rename component, make if more readable
